delete dataset objects only if they exist

+++ server/dataset.go
@@ -219,7 +219,15 @@ func (endpoint *DatasetEndpoints) DeleteDataset(ctx context.Context, request *se
 		return nil, err
 	}
 
-	err = endpoint.ObjectHandler.DeleteObjects(objects)
+	if len(objects) > 0 {
+		err = endpoint.ObjectHandler.DeleteObjects(objects)
+		if err != nil {
+			log.Println(err.Error())
+			return nil, err
+		}
+	}
+
+	err = endpoint.DeleteHandler.DeleteDataset(uint(request.GetId()))
 	if err != nil {
 		log.Println(err.Error())
 		return nil, err
